+++ backend/package.json
@@ -8,6 +8,7 @@
 		"build": "tsc"
 	},
 	"dependencies": {
+		"@aws-sdk/client-s3": "3.677.0",
 		"@bcgov/citz-imb-express-utilities": "1.0.0",
 		"@bcgov/citz-imb-sso-js-core": "1.0.0",
 		"amqplib": "0.10.4",
