+++ backend/src/modules/auth/controllers/loginCallback.ts
@@ -8,49 +8,41 @@ const { SSO_ENVIRONMENT, SSO_REALM, SSO_PROTOCOL, SSO_CLIENT_ID, SSO_CLIENT_SECR
 	ENV;
 
 export const loginCallback = errorWrapper(async (req: Request, res: Response) => {
-	try {
-		const { code } = req.query;
+	const { code } = req.query;
 
-		const tokens = await getTokens({
-			code: code as string,
-			clientID: SSO_CLIENT_ID,
-			clientSecret: SSO_CLIENT_SECRET,
-			redirectURI: `${BACKEND_URL}/auth/login/callback`,
-			ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
-			ssoRealm: SSO_REALM,
-			ssoProtocol: SSO_PROTOCOL as SSOProtocol,
-		});
+	const tokens = await getTokens({
+		code: code as string,
+		clientID: SSO_CLIENT_ID,
+		clientSecret: SSO_CLIENT_SECRET,
+		redirectURI: `${BACKEND_URL}/auth/login/callback`,
+		ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
+		ssoRealm: SSO_REALM,
+		ssoProtocol: SSO_PROTOCOL as SSOProtocol,
+	});
 
-		// Sets tokens
-		res
-			.cookie("refresh_token", tokens.refresh_token, {
-				httpOnly: true,
-				secure: true,
-				sameSite: "none",
-			})
-			.cookie("access_token", tokens.access_token, {
-				secure: true,
-				sameSite: "none",
-			})
-			.cookie("id_token", tokens.id_token, {
-				secure: true,
-				sameSite: "none",
-			})
-			.cookie("expires_in", tokens.expires_in, {
-				secure: true,
-				sameSite: "none",
-			})
-			.cookie("refresh_expires_in", tokens.refresh_expires_in, {
-				secure: true,
-				sameSite: "none",
-			})
-			.status(200)
-			.json(tokens);
-	} catch (error) {
-		res.status(500).json({
-			success: false,
-			error:
-				error instanceof Error ? error.message : "An unknown error occurred during login callback.",
-		});
-	}
+	// Sets tokens
+	res
+		.cookie("refresh_token", tokens.refresh_token, {
+			httpOnly: true,
+			secure: true,
+			sameSite: "none",
+		})
+		.cookie("access_token", tokens.access_token, {
+			secure: true,
+			sameSite: "none",
+		})
+		.cookie("id_token", tokens.id_token, {
+			secure: true,
+			sameSite: "none",
+		})
+		.cookie("expires_in", tokens.expires_in, {
+			secure: true,
+			sameSite: "none",
+		})
+		.cookie("refresh_expires_in", tokens.refresh_expires_in, {
+			secure: true,
+			sameSite: "none",
+		})
+		.status(200)
+		.json(tokens);
 });
