+++ backend/src/modules/auth/controllers/logout.ts
@@ -7,22 +7,15 @@ import { errorWrapper } from "@bcgov/citz-imb-express-utilities";
 const { SSO_ENVIRONMENT, SSO_REALM, SSO_PROTOCOL, BACKEND_URL } = ENV;
 
 export const logout = errorWrapper(async (req: Request, res: Response) => {
-	try {
-		const { id_token } = req.query;
+	const { id_token } = req.query;
 
-		const redirectURL = getLogoutURL({
-			idToken: id_token as string,
-			postLogoutRedirectURI: `${BACKEND_URL}/auth/logout/callback`,
-			ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
-			ssoProtocol: SSO_PROTOCOL as SSOProtocol,
-			ssoRealm: SSO_REALM,
-		});
+	const redirectURL = getLogoutURL({
+		idToken: id_token as string,
+		postLogoutRedirectURI: `${BACKEND_URL}/auth/logout/callback`,
+		ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
+		ssoProtocol: SSO_PROTOCOL as SSOProtocol,
+		ssoRealm: SSO_REALM,
+	});
 
-		res.redirect(redirectURL);
-	} catch (error) {
-		res.status(500).json({
-			success: false,
-			error: error instanceof Error ? error.message : "An unknown error occurred during logout.",
-		});
-	}
+	res.redirect(redirectURL);
 });
