+++ backend/src/modules/auth/controllers/token.ts
@@ -7,41 +7,31 @@ import { errorWrapper } from "@bcgov/citz-imb-express-utilities";
 const { SSO_ENVIRONMENT, SSO_REALM, SSO_PROTOCOL, SSO_CLIENT_ID, SSO_CLIENT_SECRET } = ENV;
 
 export const token = errorWrapper(async (req: Request, res: Response) => {
-	try {
-		const refresh_token = req.cookies.refresh_token;
+	const refresh_token = req.cookies.refresh_token;
 
-		if (!refresh_token)
-			return res.status(401).json({
-				success: false,
-				message: "Refresh token is missing. Please log in again.",
-			});
-
-		const tokens = await getNewTokens({
-			refreshToken: refresh_token as string,
-			clientID: SSO_CLIENT_ID,
-			clientSecret: SSO_CLIENT_SECRET,
-			ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
-			ssoRealm: SSO_REALM,
-			ssoProtocol: SSO_PROTOCOL as SSOProtocol,
-		});
-
-		if (!tokens) return res.status(401).json({ success: false, message: "Invalid token." });
-
-		// Set token
-		res
-			.cookie("access_token", tokens.access_token, {
-				secure: true,
-				sameSite: "none",
-			})
-			.status(200)
-			.json(tokens);
-	} catch (error) {
-		res.status(500).json({
+	if (!refresh_token)
+		return res.status(401).json({
 			success: false,
-			error:
-				error instanceof Error
-					? error.message
-					: "An unknown error occurred while refreshing tokens.",
+			message: "Refresh token is missing. Please log in again.",
 		});
-	}
+
+	const tokens = await getNewTokens({
+		refreshToken: refresh_token as string,
+		clientID: SSO_CLIENT_ID,
+		clientSecret: SSO_CLIENT_SECRET,
+		ssoEnvironment: SSO_ENVIRONMENT as SSOEnvironment,
+		ssoRealm: SSO_REALM,
+		ssoProtocol: SSO_PROTOCOL as SSOProtocol,
+	});
+
+	if (!tokens) return res.status(401).json({ success: false, message: "Invalid token." });
+
+	// Set token
+	res
+		.cookie("access_token", tokens.access_token, {
+			secure: true,
+			sameSite: "none",
+		})
+		.status(200)
+		.json(tokens);
 });
