+++ backend/src/modules/rabbit/controllers/index.ts
@@ -1,1 +1,1 @@
-export * from "./addToRabbitQueue";
+export * from "./addToTestQueue";
