+++ backend/src/modules/s3/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from "./health";
