+++ backend/src/modules/s3/index.ts
@@ -1,0 +1,1 @@
+export { default as router } from "./router";
