+++ backend/src/modules/s3/router.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { health } from "./controllers";
+
+const router = Router();
+
+router.post("/health", health);
+
+export default router;
