+++ backend/tests/modules/rabbit/router.test.ts
@@ -2,11 +2,11 @@ import type { Request, Response } from "express";
 import express from "express";
 import request from "supertest";
 import router from "@/modules/rabbit/router";
-import { addToRabbitQueue } from "@/modules/rabbit/controllers";
+import { addToTestQueue } from "@/modules/rabbit/controllers";
 
 // Mock the controller functions
 jest.mock("@/modules/rabbit/controllers", () => ({
-	addToRabbitQueue: jest.fn((req: Request, res: Response) => res.status(200).send("Complete")),
+	addToTestQueue: jest.fn((req: Request, res: Response) => res.status(200).send("Complete")),
 }));
 
 // Helper function to create an Express app with the router
@@ -17,10 +17,10 @@ const createApp = () => {
 };
 
 describe("Rabbit Router", () => {
-	// Test case for /addToQueue route
-	it("should call the addToRabbitQueue controller on /addToQueue route", async () => {
+	// Test case for /addToTestQueue route
+	it("should call the addToTestQueue controller on /addToTestQueue route", async () => {
 		const app = createApp();
-		await request(app).post("/addToQueue").expect(200, "Complete");
-		expect(addToRabbitQueue).toHaveBeenCalled();
+		await request(app).post("/addToTestQueue").expect(200, "Complete");
+		expect(addToTestQueue).toHaveBeenCalled();
 	});
 });
